fixed authentication failure

+++ lib/vws.rb
@@ -143,24 +143,25 @@ module Vws
       #active_flag to false. Then we post a delete action to delete the
       #target
       target_info = JSON.parse(retrieve_target(target_id))
-      if target_info["target_record"]["active_flag"] = "true" &&  
-        target_info["target_record"]["status"] != "processing" 
-        self.set_active_flag(target_id, false)
-      elsif target_info["target_record"]["active_flag"] = "false" && 
-        target_info["target_record"]["status"] == "success"
-        begin
-          date_timestamp = Time.now.httpdate
-          target_id_url = TARGETS_URL + '/' + target_id
-          target_id_suburl = '/targets' + '/' + target_id
-          signature = self.build_signature(target_id_suburl, nil, 'DELETE', date_timestamp)
-          authorization_header = "VWS " + @accesskey + ":" + signature
-          RestClient.delete(target_id_url, :'Date' => date_timestamp,
+      if target_info["result_code"] == "Success"
+        if target_info["target_record"]["active_flag"] = "true" &&  
+          target_info["target_record"]["status"] != "processing" 
+          self.set_active_flag(target_id, false)
+        elsif target_info["target_record"]["active_flag"] = "false" && 
+          target_info["target_record"]["status"] == "success"
+          begin
+            date_timestamp = Time.now.httpdate
+            target_id_url = TARGETS_URL + '/' + target_id
+            target_id_suburl = '/targets' + '/' + target_id
+            signature = self.build_signature(target_id_suburl, nil, 'DELETE', date_timestamp)
+            authorization_header = "VWS " + @accesskey + ":" + signature
+            RestClient.delete(target_id_url, :'Date' => date_timestamp,
                                              :'Authorization' => authorization_header)        
-        rescue => e
-          e.response
+          rescue => e
+            e.response
+          end
         end
+       end
       end
-    end
-
   end
 end
